requirements.txt
torch~=1.8.1
numpy~=1.19.2
pillow~=8.2.0
pandas~=1.2.4
tensorflow~=1.14.0
seaborn~=0.11.1
matplotlib~=3.4.2
scikit-learn~=0.24.2
scikit-image~=0.18.1
graphviz==0.19.1
scipy~=1.6.2
torchvision~=0.9.1
neptune-client==0.9.19